update signup, login

+++ shopapp-backend/shopapp/src/main/java/com/project/shopapp/dtos/OrderDetailDTO.java
@@ -25,7 +25,7 @@ public class OrderDetailDTO {
     private Long productId;
 
     @Min(value = 0, message = "Price must be >=0")
-    private Long price;
+    private Float price;
 
     @JsonProperty("number_of_products")
     @Min(value = 1, message = "Price must be >=1")
@@ -33,7 +33,9 @@ public class OrderDetailDTO {
 
     @JsonProperty("total_money")
     @Min(value = 0, message = "Total money must be >= 0")
-    private Integer totalMoney;
+    private Float totalMoney;
 
     private String color;
+
+
 }
